UniAD/requirements2.0_rocm.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
backcall==0.2.0
black==24.8.0
cachetools==5.5.1
casadi==3.6.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
cmake==3.25.0
cycler==0.12.1
decorator==5.1.1
descartes==1.1.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.13.1
fire==0.7.0
flake8==7.1.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
google-api-core==2.24.1
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.4
imageio==2.35.1
importlib_metadata==8.5.0
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lightning-utilities==0.11.9
lit
llvmlite
lyft-dataset-sdk==0.0.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mccabe==0.7.0
motmetrics==1.4.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.26.0
networkx==3.4.2
numba
numpy==1.22.4
nuscenes-devkit==1.1.11
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
plyfile==1.0.3
prettytable==3.11.0
prompt_toolkit==3.0.50
propcache==0.2.0
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.1.4
pyquaternion==0.9.9
pytest==8.3.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.2.5
pytz==2025.1
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.28.1
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
Shapely==1.8.5.post1
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
tomli==2.2.1
# torch==2.0.1+cu118
# torchaudio==2.0.2+cu118
# torchmetrics==1.5.2
# torchvision==0.15.2+cu118
tqdm==4.67.1
traitlets==5.14.3
trimesh==2.35.39
triton==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.13
wcwidth==0.2.13
Werkzeug==3.0.6
xmltodict==0.14.2
yapf==0.40.1
yarl==1.15.2
zipp==3.20.2
